backend/utils/jwt_handler.py
import jwt
import datetime
from config.config import JWT_SECRET_KEY

def create_jwt_token(payload: dict) -> str:
    """Create a JWT token"""
    # Add expiration time (24 hours from now)
    payload['exp'] = datetime.datetime.utcnow() + datetime.timedelta(days=1)
    return jwt.encode(payload, JWT_SECRET_KEY, algorithm='HS256')

def verify_jwt_token(token: str) -> dict:
    """Verify a JWT token and return its payload"""
    try:
        return jwt.decode(token, JWT_SECRET_KEY, algorithms=['HS256'])
    except jwt.ExpiredSignatureError:
        raise ValueError('Token has expired')
    except jwt.InvalidTokenError:
        raise ValueError('Invalid token')
