backend/requirements.txt
Flask==2.0.1
Werkzeug==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
requests==2.28.2
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
python-multipart==0.0.6
PyJWT==2.8.0
